🔒️SECURITY:  make jwt session short-live (5 min)
This commit change the lifetime to 5 min (as seen in session_setting.ts),
In addtion it set keepAlive and restructure jwt flow to be more efficent.

jwt session is create on signIn and stored on client side.
It's signed by server so it can be use as credential which reduce the need to fetch credential from database.
By default next-auth have jwt lifetime/age of 30 day
which mean if someone have jwt credential (say, being an admin) they can be admin for 30 day after thier admin credential get revoked if they didn't logout.